Fix #575. Add x64 configurations to some key foundation components.
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@3185 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Common/Foundation/System/StreamReader.cpp
@@ -459,10 +459,10 @@ MgByteReader* MgStreamReader::GetStream(bool enableDirect)
                 }
                 else
                 {
-                    UINT32 dataWritten = 0;
+                    ssize_t dataWritten = 0;
                     while (dataWritten < len)
                     {
-                        int written = ACE_OS::write(file, &temp[dataWritten], len-dataWritten);
+                        ssize_t written = ACE_OS::write(file, &temp[dataWritten], len-dataWritten);
                         if (written > 0)
                         {
                             dataWritten += written;
