+++ Oem/ACE/ACE_wrappers/ace/OS_NS_unistd.h
@@ -238,12 +238,12 @@ namespace ACE_OS {
                   size_t *bytes_transferred = 0);
 
   ACE_NAMESPACE_INLINE_FUNCTION
-  int readlink (const char *path,
+  ssize_t readlink (const char *path,
                 char *buf,
                 size_t bufsiz);
 
   ACE_NAMESPACE_INLINE_FUNCTION
-  void *sbrk (int brk);
+  void *sbrk (intptr_t brk);
 
   ACE_NAMESPACE_INLINE_FUNCTION
   int setgid (gid_t);
