+++ Oem/dbxml-2.3.10/xerces-c-src/src/xercesc/internal/XMLScanner.hpp
@@ -1289,7 +1289,7 @@ inline void XMLScanner::setValidateAnnotations(const bool newValue)
 inline void XMLScanner::setInputBufferSize(const size_t bufferSize)
 {
     fBufferSize = bufferSize;
-    fCDataBuf.setFullHandler(this, fBufferSize);
+    fCDataBuf.setFullHandler(this, (unsigned int)fBufferSize);
 }
 
 inline void XMLScanner::setIgnoredCachedDTD(const bool newValue)
