+++ Oem/dbxml-2.3.10/xqilla/include/xqilla/ast/XQNav.hpp
@@ -66,7 +66,7 @@ protected:
 class XQILLA_API NavStepResult : public ResultImpl
 {
 public:
-  NavStepResult(const Result &parent, ASTNode *step, unsigned int contextSize);
+  NavStepResult(const Result &parent, ASTNode *step, size_t contextSize);
 
   Item::Ptr next(DynamicContext *context);
   std::string asString(DynamicContext *context, int indent) const;
@@ -76,8 +76,8 @@ private:
   Result parent_;
   const ASTNode *step_;
   Result stepResult_;
-  unsigned int contextPos_;
-  unsigned int contextSize_;
+  size_t contextPos_;
+  size_t contextSize_;
   Item::Ptr contextItem_;
 };
