+++ Oem/dbxml-2.3.10/xqilla/include/xqilla/context/impl/XQDynamicContextImpl.hpp
@@ -91,14 +91,14 @@ public:
   virtual void setExternalContextNode(const XERCES_CPP_NAMESPACE_QUALIFIER DOMNode *node);
 
   /** Get the context position */
-  virtual unsigned int getContextPosition() const;
+  virtual size_t getContextPosition() const;
   /** Set the context position */
-  virtual void setContextPosition(unsigned int pos);
+  virtual void setContextPosition(size_t pos);
 
   /** Get the context size */
-  virtual unsigned int getContextSize() const;
+  virtual size_t getContextSize() const;
   /** Set the context size */
-  virtual void setContextSize(unsigned int size);
+  virtual void setContextSize(size_t size);
 
   /** get the variable store */
   virtual VariableStore* getVariableStore();
@@ -297,11 +297,11 @@ protected:
 
   /** The context position is the position of  the context item within
    * the sequence of items currently being processed. */
-  unsigned int _contextPosition;
+  size_t _contextPosition;
 
   /** The context size is the number of items in the sequence of
    * items currently being processed. */
-  unsigned int _contextSize;
+  size_t _contextSize;
 
   /** In-scope variable values. This is a set of  (QName, Sequence) pairs.
    * It defines the set of variables that have been declared and are
